Deleted hub usage Added exception handling Added request logging
