--- src/main/java/com/windriver/core/ErrorHandler.java
@@ -1,19 +1,0 @@
-package com.windriver.core;
-
-import com.windriver.util.InputStreamConverter;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.client.ClientHttpResponse;
-import org.springframework.web.client.DefaultResponseErrorHandler;
-
-import java.io.IOException;
-
-public class ErrorHandler extends DefaultResponseErrorHandler {
-
-    private static final Logger log = LoggerFactory.getLogger(ErrorHandler.class);
-
-    @Override
-    public void handleError(ClientHttpResponse response) throws IOException {
-        log.error("Error on agent side: " + InputStreamConverter.convert(response.getBody()));
-    }
-}
