--- src/main/java/com/windriver/model/automation/FindOption.java
@@ -1,17 +1,0 @@
-package com.windriver.model.automation;
-
-public enum FindOption implements WinDriverOption {
-
-    FIND_FIRST("FindFirst"),
-    FIND_ALL("FindAll");
-
-    private String option;
-
-    FindOption(String option) {
-        this.option = option;
-    }
-
-    public String get() {
-        return option;
-    }
-}
