--- src/main/java/com/windriver/model/automation/WinDriverOption.java
@@ -1,9 +1,0 @@
-package com.windriver.model.automation;
-
-import com.fasterxml.jackson.annotation.JsonValue;
-
-public interface WinDriverOption {
-
-    @JsonValue
-    public String get();
-}
