--- src/main/java/com/windriver/util/InputStreamConverter.java
@@ -1,11 +1,0 @@
-package com.windriver.util;
-
-import java.io.InputStream;
-import java.util.Scanner;
-
-public class InputStreamConverter {
-
-    public static String convert(InputStream inputStream) {
-        return new Scanner(inputStream, "UTF-8").useDelimiter("\\A").next();
-    }
-}
