+++ src/main/java/ua/windriver/model/automation/WinDriverOption.java
@@ -1,0 +1,9 @@
+package ua.windriver.model.automation;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public interface WinDriverOption {
+
+    @JsonValue
+    public String get();
+}
