+++ src/main/java/ua/windriver/model/exception/WinDriverAgentException.java
@@ -1,0 +1,10 @@
+package ua.windriver.model.exception;
+
+public class WinDriverAgentException extends RuntimeException {
+
+    public WinDriverAgentException(String msg) {
+        super(msg);
+    }
+    public WinDriverAgentException() {
+    }
+}
