+++ src/main/java/ua/windriver/model/exception/WinDriverElementNotFoundException.java
@@ -1,0 +1,10 @@
+package ua.windriver.model.exception;
+
+public class WinDriverElementNotFoundException extends RuntimeException {
+
+    public WinDriverElementNotFoundException(String msg) {
+        super(msg);
+    }
+    public WinDriverElementNotFoundException() {
+    }
+}
