+++ src/main/java/ua/windriver/model/response/ActionControlResponse.java
@@ -1,0 +1,49 @@
+package ua.windriver.model.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ActionControlResponse {
+
+    @JsonProperty("Message")
+    private String message;
+
+    @JsonProperty("Entity")
+    private Object entity;
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public Object getEntity() {
+        return entity;
+    }
+
+    public void setEntity(Object entity) {
+        this.entity = entity;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        ActionControlResponse that = (ActionControlResponse) o;
+
+        if (message != null ? !message.equals(that.message) : that.message != null) return false;
+        return entity != null ? entity.equals(that.entity) : that.entity == null;
+
+    }
+
+    @Override
+    public int hashCode() {
+        int result = message != null ? message.hashCode() : 0;
+        result = 31 * result + (entity != null ? entity.hashCode() : 0);
+        return result;
+    }
+}
