+++ src/main/resources/windriver.properties
@@ -1,15 +1,9 @@
-#hub properties
-hub.url.connection.type=http
-hub.url.host=localhost
-hub.url.port=8079
+#agent properties
+agent.url.connection.type=http
+agent.url.host=10.25.13.107
+agent.url.port=8082
 
 #client properties
 client.log.level=INFO
 http.request.read.timeout=20000
 http.request.connect.timeout=20000
-
-#local agent properties
-agent.url.connection.type=http
-agent.url.host=localhost
-agent.url.port.min=8000
-agent.url.port.max=8050
